requirements.txt
imgaug==0.4.0
matplotlib==3.3.4
numpy==1.20.1
opencv_contrib_python==4.5.4.58
Pillow==9.2.0
PyYAML==6.0
scikit_learn==1.1.1
scipy==1.6.2
torch==1.8.2+cu111
torchvision==0.9.2+cu111
tqdm==4.59.0
